Move transformArticle into its own file
